Move spatial models to their own file
