+++ src/opencosmo/collection/io.py
@@ -1,11 +1,13 @@
 from pathlib import Path
-from typing import Type
+from typing import TYPE_CHECKING, Type
 
 import opencosmo as oc
-from opencosmo.collection.protocols import Collection
 from opencosmo.collection.simulation import SimulationCollection
 from opencosmo.collection.structure.io import validate_linked_groups
 from opencosmo.io.io import FILE_TYPE, OpenTarget
+
+if TYPE_CHECKING:
+    from opencosmo.collection.protocols import Collection
 
 
 def open_simulation_files(**paths: Path) -> SimulationCollection:
@@ -36,7 +38,7 @@ def open_simulation_files(**paths: Path) -> SimulationCollection:
 
 def get_collection_type(
     targets: list[OpenTarget], file_types: list[FILE_TYPE]
-) -> Type[Collection]:
+) -> Type["Collection"]:
     """
     If there are multiple files, determine their collection type. There are
     three options we support at present:
